refactor: extract skill file content normalization
Move text-extension policy, UTF-16 detection, and byte-level line-ending normalization from SkillInstallLayout into a focused SkillFileContentNormalizer.

Update the two production callers and focused tests to reference the extracted class directly, leaving no compatibility wrapper or behavior change.
